% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_helpers.R
\name{as.data.frame.knnst}
\alias{as.data.frame.knnst}
\title{Convert a \code{knnst} object to a \code{data.frame}}
\usage{
\method{as.data.frame}{knnst}(x, ...)
}
\arguments{
\item{x}{A \code{knnst} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
\code{knnst} objects can be converted to \code{data.frame}s. When doing so, the
disaggregated monthly data are combined with the simulation number, and
the index years, are repeated for each month. The rownames of the
disaggregated data (yyyy-mm) format are kept, and additional columns are
added for year and month, by themselves.
}
