% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_disagg}
\alias{ex_disagg}
\title{Example disaggregated data}
\format{
A \code{\link{knnst}} object.
}
\usage{
ex_disagg
}
\description{
An example \code{\link{knnst}} object containing results of disaggregating the first
200 years in the \code{\link{meko}} data. Data were disaggregated using the
\href{https://github.com/BoulderCodeHub/CoRiverNF}{CoRiverNF} package for the
historical annual data and monthly pattern data. Including this object as it
is useful for using in examples and tests.
}
\references{
See
\href{https://github.com/rabutler-usbr/knnstdisagg/data-raw}{data-raw} for code
used to generate data.
}
\keyword{datasets}
