% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_params.R
\name{is_knn_params}
\alias{is_knn_params}
\title{Test if the object inherits from  \code{knn_params} class}
\usage{
is_knn_params(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{knn_params} class.
}
\description{
Test if the object inherits from  \code{knn_params} class
}
