% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_helpers.R
\name{is_knnst}
\alias{is_knnst}
\title{Test if the object inherits from  \code{knnst} class}
\usage{
is_knnst(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{knnst} class.
}
\description{
Test if the object inherits from  \code{knnst} class
}
