% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_params.R
\name{knn_params}
\alias{knn_params}
\title{Class to specify the parameters for the knn selection}
\usage{
knn_params(k, weights)
}
\arguments{
\item{k}{A single integer specifying the number of neighbors to
select from.}

\item{weights}{a vector with length equal to \code{k}, specifying the weights to
apply to the 1-\code{k} neighbors before selecting the neighbor. The weights
must sum to 1.}
}
\description{
\code{knn_params()} creates an object, which lets the user specify values for
the weights, and the number of neighbors to use when selecting the
nearest neighbor.
}
\examples{
# uniformly select from the 5 nearest neighbors
knn_params(5, rep(1/5, 5))
# select from 4 neigbors, with decreasing likelihood of being selected
knn_params(4, c(.4, .3, .2, .1))
# select the nearest neighbor
knn_params(1, 1)

}
