% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_disagg.R
\name{knn_space_time_disagg}
\alias{knn_space_time_disagg}
\title{Spatial and temporal diaggregattion of paleo flow data}
\usage{
knn_space_time_disagg(x, ann_index_flow, mon_flow, nsim = 1,
  sf_sites = NULL, ofolder = NULL, index_years = NULL, k_weights = NULL)
}
\arguments{
\item{x}{The annual paleo data to disaagregate.}

\item{ann_index_flow}{Observed annual flow data used for picking index year.}

\item{mon_flow}{Intervening monthly natural flow. Used for spatially and
temporaly disaggregating paleo data based on the index year. Each column
should be a differnt site to disaggregate the flow to. If there are three
columns in this matrix, then the values in \code{x} will be disaggregated to
three sites. \code{mon_flow} should have the same years as \code{ann_index_flow}}

\item{nsim}{Number of times to repeat the space/time disaggregation.}

\item{sf_sites}{The site numbers (column indeces), that will scale the
index year's volume based on the annual flow being disaggregated. The
remaining sites will select the index year directly. See \strong{Details}.}

\item{ofolder}{Optional. If specified, the disaggregated flow data and the
selected index years are saved to this folder as csv files. There will be
one csv file for each time the disaggregation is repeated (\code{nsim}). This
file will contain one column for each site (\code{nsite}), and one row for each
month of data (12 * number of years in \code{x}).}

\item{index_years}{Optional. If specified, these index years will be used
instead of selecting years based on weights and sampling.}

\item{k_weights}{If \code{NULL}, parameters are set based on definitions in Nowak
et al. (2010). Users may force \code{k} and the \code{weights} by specifiying this
argument. It should be a list with two named entries: \code{k} and \code{weights}.}
}
\description{
The default parameter values are setup to perform the typical disaggregation
for CRSS, based on scaling the Upper Basin inflows, and performing a
direct resampling of Lower Basin tributaries, i.e., no scale factor is
applied to the LB tributaries. Sites 1-20 are scaled (\code{sf_sites = 1:20});
therefore the remaining sites (21:29) are not scaled.
}
\examples{
\dontrun{
# read in annual synthetic mon_flow for disag
x <- matrix(scan("data-raw/Meko.txt"), ncol = 2, byrow = TRUE)
# intervening natural flow mon_flow - monthly CY text file
mon_flow <- as.matrix(read.table(
  "tests/dp/NFfullbasinWY0608intervening.txt",
  sep = "\\t"
))

# observed annual flow for picking analog disag yr
ann_index_flow <- as.matrix(read.table("tests/dp/LFWYTotal.txt"))
zz <- paleo_disagg(x, ann_index_flow, mon_flow, 29, 1)
}

}
\references{
Nowak, K., Prairie, J., Rajagopalan, B., Lall, U. (2010).
A nonparametric stochastic approach for multisite disaggregation of
annual to daily streamflow. \emph{Water Resources Research.}
}
\author{
Ken Nowak
}
