% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_helpers.R
\name{knnst_get_disagg_data}
\alias{knnst_get_disagg_data}
\title{Get the disaggregated data}
\usage{
knnst_get_disagg_data(disagg, sim_num = 1)
}
\arguments{
\item{disagg}{A \code{knnst} object}

\item{sim_num}{The simulation number.}
}
\value{
A matrix.
}
\description{
\code{knnst_get_disagg_data()} gets the disaggregated streamflow data from a
\code{knnst} object.
}
\details{
As \code{knnst} objects can contain multiple simulations of disaggregated data,
\code{sim_num} specifies which simulation of data to return.
}
