% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_helpers.R
\name{knnst_nsim}
\alias{knnst_nsim}
\title{Get the number of diaggregation simulations from \code{knnst} objects}
\usage{
knnst_nsim(disagg)
}
\arguments{
\item{disagg}{A \code{knnst} object}
}
\value{
The number of diaggregation simulations
}
\description{
Get the number of diaggregation simulations from \code{knnst} objects
}
\examples{
flow_mat <- cbind(c(2000, 2001, 2002), c(1400, 1567, 1325))
# made up historical data to use as index years
ind_flow <- cbind(1901:1980, rnorm(80, mean = 1500, sd = 300))
# make up monthly flow for two sites
mon_flow <- cbind(
  rnorm(80 * 12, mean = 20, sd = 5),
  rnorm(80 * 12, mean = 120, sd = 45)
)
x <- knn_space_time_disagg(
  flow_mat, ind_flow, mon_flow, sf_sites = 1:2, nsim = 5
)

# will return 5
knnst_nsim(x)

}
