% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_sp_cor.R
\name{knnst_sp_cor}
\alias{knnst_sp_cor}
\title{Computes spatial correlation after knn space-time disaggregation}
\usage{
knnst_sp_cor(disagg, sites, bin_size)
}
\arguments{
\item{disagg}{A \code{knnst} object}

\item{sites}{One or more sites to compute correlations from.}

\item{bin_size}{Number of years for each bin.}
}
\value{
A \code{knnst_spcor} object.
}
\description{
Computes the spatial correlation for all monthly data between the specifed
site(s) (\code{sites}) and all other sites.
}
\examples{
knnst_sp_cor(ex_disagg, "Cameo", 100)
knnst_sp_cor(ex_disagg, c("Cameo", "Hoover"), 100)

}
