% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_spatial_cor.R
\name{knnst_spatial_cor}
\alias{knnst_spatial_cor}
\title{Compute spatial correlation after KNN space-time disaggregation}
\usage{
knnst_spatial_cor(disagg, sites, bin_size)
}
\arguments{
\item{disagg}{A \code{knnst} object}

\item{sites}{One or more sites to compute correlations from.}

\item{bin_size}{Number of years for each bin.}
}
\value{
A \code{knnst_spcor} object.
}
\description{
\code{knnst_spatial_cor()} computes the spatial correlation for all monthly data
between the specifed site(s) (\code{sites}) and all other sites.
}
\examples{
knnst_spatial_cor(ex_disagg, "Cameo", 100)
knnst_spatial_cor(ex_disagg, c("Cameo", "Hoover"), 100)

}
\seealso{
\code{\link[=plot.knnst_spcor]{plot.knnst_spcor()}}, \code{\link[=knnst_temporal_cor]{knnst_temporal_cor()}}
}
