% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnst_temporal_cor.R
\name{knnst_temporal_cor}
\alias{knnst_temporal_cor}
\title{Compute temporal correlation after KNN space-time disaggregation}
\usage{
knnst_temporal_cor(disagg, site, bin_size = NULL)
}
\arguments{
\item{disagg}{A \code{knnst} object}

\item{site}{Site to compute the correlations at. Scalar character.}

\item{bin_size}{Number of years for each bin when there is only one
simulation.}
}
\value{
A \code{knnst_tmpcor} object.
}
\description{
\code{knnst_temporal_cor()} computes the temporal correlation between all months
at the specified \code{site}
}
\details{
If there are more  than one simulations, the correlation is computed for each
simulation across all years. If there is only one simulation, then the user
must specify the \code{bin_size} that is used to compute the correlation across
moving windows of this specified size.
}
\examples{
# use a 50-year bin size to compute the stats at Cameo
tmp_cor <- knnst_temporal_cor(ex_disagg, "Cameo", 50)

# use a 100-year bin size for computing the stats at Hoover
tmp_cor <- knnst_temporal_cor(ex_disagg, "Hoover", 100)

}
\seealso{
\code{\link[=plot.knnst_tmpcor]{plot.knnst_tmpcor()}}, \code{\link[=knnst_spatial_cor]{knnst_spatial_cor()}}
}
