% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meko}
\alias{meko}
\title{Upper Colorado River Flow Reconstruction.}
\format{
A matrix with 1244 rows and 2 columns:
\itemize{
\item column 1 = water_year: the water year (October - September)
\item column 2 = lees_ferry: reconstructed flow at Lees Ferry in acre-feet
}
}
\source{
Description of data:
\url{https://www.treeflow.info/content/colorado-r-lees-ferry-az-meko}

Data download:
\url{http://www.riversimulator.org/Resources/ClimateData/Meko762-2005.xls}
}
\usage{
meko
}
\description{
A dataset containing paleo reconstructed hydrology for the Colorado River at
Lees Ferry for 762 - 2005.
}
\references{
Meko, D.M., Woodhouse, C.A., Baisan, C.A., Knight, T., Lukas,
J.J., Hughes, M.K., and Salzer, M.W. 2007. Medieval Drought in the Upper
Colorado River Basin. Geophysical Research Letters 34, L10705.
}
\keyword{datasets}
