% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_disagg.R
\name{paleo_disagg}
\alias{paleo_disagg}
\title{Spatial and temporal diaggregattion of paleo flow data}
\usage{
paleo_disagg(x, ann_index_flow, mon_flw, nsite = 29, sf_sites = 1:20,
  nsim = 1, ofolder = NULL, index_years = NULL, k_weights = NULL)
}
\arguments{
\item{x}{The annual paleo data to disaagregate.}

\item{ann_index_flow}{Observed annual flow data used for picking index year.}

\item{mon_flw}{Intervening monthly natural flow. Used for spatially and
temporaly disaggregating paleo data based on the index year.}

\item{nsite}{The number of sites to diaggregate the data too.}

\item{sf_sites}{The site numbers (indeces), that will scale the index year's
volume based on the annual flow being disaggregated. The remaining sites
will select the index year directly. See **Details**.}

\item{nsim}{Number of times to repeat the space/time disaggregation.}

\item{ofolder}{Optional. If specified, the disaggregated flow data and the
selected index years are saved to this folder as csv files. There will be
one csv file for each time the disaggregation is repeated (`nsim`). This
file will contain one column for each site (`nsite`), and one row for each
month of data (12 * number of years in `x`).}

\item{index_years}{Optional. If specified, these index years will be used
instead of selecting years based on weights and sampling.}

\item{k_weights}{If `NULL`, parameters are set based on definitions in Nowak
et al. (2010). Users may force `k` and the `weights` by specifiying this
argument. It should be a list with two named entries: `k` and `weights`.}
}
\description{
The default parameter values are setup to perform the typical disaggregation
for CRSS, based on scaling the Upper Basin inflows, and performing a
direct resampling of Lower Basin tributaries, i.e., no scale factor is
applied to the LB tributaries. Sites 1-20 are scaled (`sf_sites = 1:20`);
therefore the remaining sites (21:29) are not scaled.
}
\examples{
\dontrun{
# read in annual synthetic mon_flw for disag
x <- matrix(scan("data-raw/Meko.txt"), ncol = 2, byrow = TRUE)
# intervening natural flow mon_flw - monthly CY text file
mon_flw <- as.matrix(read.table(
  "tests/dp/NFfullbasinWY0608intervening.txt",
  sep = "\\t"
))

# observed annual flow for picking analog disag yr
ann_index_flow <- as.matrix(read.table("tests/dp/LFWYTotal.txt"))
zz <- paleo_disagg(x, ann_index_flow, mon_flw, 29, 1)
}

}
\references{
Nowak, K., Prairie, J., Rajagopalan, B., Lall, U. (2010).
  A nonparametric stochastic approach for multisite disaggregation of
  annual to daily streamflow. *Water Resources Research.*
}
\author{
Ken Nowak
}
