% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knnst.R
\name{plot.knnst}
\alias{plot.knnst}
\title{Plot resulting statistics for knn space-time disaggregation}
\usage{
\method{plot}{knnst}(
  x,
  site,
  bin_size = NULL,
  base_units = NULL,
  which = c(13, 14, 15),
  show = FALSE,
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{knnst}}

\item{site}{The site to plot. Site name as a character.}

\item{bin_size}{Number of years for each bin when there is only one
simulation.}

\item{base_units}{The (input) units of the flow that was disaggregated. These
units will be shown as the y-axis label of the plot.}

\item{which}{The subset of plots to create; specify a subset of the number
\code{1:15}. See 'Details' for the different plots.}

\item{show}{Boolean that determines if diagnostic plots are shown or only
returned. If \code{TRUE} and in interactive mode, will be able to enter through
each of the 1-15 plots.}

\item{...}{Arguments to be passed to subsequent methods.}
}
\value{
A \code{knnstplot} object
}
\description{
\code{plot.knnst()} implements the \code{\link[=plot]{plot()}} method for \code{knnst} objects, and relies
on ggplot2. It plots the key statistics the knn space-time disaggregation
method should be preserving.
}
\details{
If there are more  than one simulations, the stats are computed for each
simulation across all years. If there is only one simulation, then the user
must specify the \code{bin_size} that is used to compute the stats across
moving windows of this specified size.

\code{which} controls the plots that are created. There are both monthly and
annual plots and two main plot types: a cdf of the flows across all of the
different disaggregations, and a panel showing the mean, max, min, variance,
lag-1 correlation, and skew across all of the different disaggregations. The
numbers correspond to:
\itemize{
\item \code{1:12}: the different monthly cdfs.
\item \code{13}: annual cdf
\item \code{14}: monthly statistics panel
\item \code{15}: annual statistics panel
}

\code{...} is passed to \code{\link[=geom_point]{geom_point()}} and \code{\link[=geom_line]{geom_line()}} to overwrite defaults for
\code{size}, \code{shape}, and \code{color}.

For the monthly statistics panel, months are ordered based on the
\code{start_month} provided in the original disaggregation, i.e., specified by the
user in \code{\link[=knn_space_time_disagg]{knn_space_time_disagg()}}.
}
