% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knnst_spcor.R
\name{plot.knnst_spcor}
\alias{plot.knnst_spcor}
\title{Plot resulting spatial correlation values after KNN space-time disaggregation}
\usage{
\method{plot}{knnst_spcor}(x, x_names = FALSE, ncol = 1, nrow = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{knnst_spcor}.}

\item{x_names}{Boolean. If \code{TRUE} use names for the x axis labels, otherwise
use site numbers.}

\item{ncol, nrow}{Number of columns and rows to use in
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{...}{Arguments to be passed to subsequent methods not used in this
method.}
}
\description{
\code{plon.knnst_spcor()} implements the \code{plot()} method for \code{knnst_spcor} objects
relying on ggplot2. See \code{\link[=knnst_spatial_cor]{knnst_spatial_cor()}} to create \code{knnst_spcor} objects.
Each site a correlation was computed from will be its own facet in the plot.
}
\examples{
zz <- knnst_spatial_cor(ex_disagg, c("Cameo", "Hoover"), 100)
# show as 1 column and 2 rows
plot(zz)

# show as 2 columns and 1 row
plot(zz, ncol = 2)

# include the site names in the x axis labels
plot(zz, x_names = TRUE)

}
