% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.knnstplot.R
\name{print.knnstplot}
\alias{print.knnstplot}
\title{Print a knnstplot}
\usage{
\method{print}{knnstplot}(x, ...)
}
\arguments{
\item{x}{An object of type \code{knnstplot}, i.e., object returned from
\code{\link[=plot.knnst]{plot.knnst()}}.}

\item{...}{other arguments not used by this method}
}
\value{
Invisibly returns \code{x}
}
\description{
Interactively prints the \code{knnstplot}. If printing in a script, use
\code{\link[=save_knnstplot]{save_knnstplot()}} instead.
}
