% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_knnstplot.R
\name{save_knnstplot}
\alias{save_knnstplot}
\title{Save a knnstplot}
\usage{
save_knnstplot(knnstplot, filename, width = NA, height = NA)
}
\arguments{
\item{knnstplot}{An object of type \code{knnstplot}, i.e., object returned from
\code{\link[=plot.knnst]{plot.knnst()}}.}

\item{filename}{File name to create on disk. Should be a .png or .pdf file.
If it is a .png file, then it will create multiple files with \verb{\%02d}
inserted at the end of the file to ensure unique file names. Directory
should already exist.}

\item{width, height}{Plot size in inches. If not specified, defaults to 7.}
}
\value{
Invisibly returns \code{knnstplot}
}
\description{
\code{save_knnstplot()} is a convenient function for saving the knn space-time
disaggregation diagnostic plots returned by \code{\link[=plot.knnst]{plot.knnst()}} as either a pdf or
collection of png files.
}
