% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_knnst.R
\name{write_knnst}
\alias{write_knnst}
\title{Write all the knnst disagg data to disk}
\usage{
write_knnst(disagg, path)
}
\arguments{
\item{disagg}{A knnst object}

\item{path}{A character scalar containing the path to the folder to write the
disaggregated data.}
}
\description{
\code{write_knnst()} writes all of the disaggregated data and the index years
selected for the disaggregation for each simulation to disk.
}
\details{
The disagggregated data and the index years are saved to seperate files.
The disaggregated data has one file for each simulation, while all index
years are saved to one file. The dissaggregated data are saved as
"disagg_[n].csv", where n is the simulation number. All index years are
saved to index_years.csv.
}
