% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_disagg.R
\name{knn_space_time_disagg}
\alias{knn_space_time_disagg}
\title{Spatial and temporal diaggregattion of flow data}
\usage{
knn_space_time_disagg(ann_flow, ann_index_flow, mon_flow, nsim = 1,
  sf_sites = NULL, ofolder = NULL, index_years = NULL,
  k_weights = knn_params_default(nrow(ann_index_flow)))
}
\arguments{
\item{ann_flow}{A 2-column matrix, with years in column 1, and an annual flow
in column 2. This is the annual flow that needs to be disaggregated.}

\item{ann_index_flow}{A 2-column matrix, with years in column 1, and an
annual flow in column 2. This is the index gage that the flows in
\code{ann_flow} will be compared to. After the comparison, the nearest neighbor
years are selected.}

\item{mon_flow}{Monthly natural flow. Used for spatially and
temporaly disaggregating the flow data (\code{ann_flow}) based on the index year
selected from \code{ann_index_flow}, by \code{\link[=knn_get_index_year]{knn_get_index_year()}}. Each column
represents a differnt site, and the annual flow at the index gage will be
disaggregated to each of these sites at he monthly level. If there are
three columns in this matrix, then the values in \code{ann_flow} will be
disaggregated to three sites. \code{mon_flow} should have the same years as
\code{ann_index_flow}, therefore, it should contain 12 times more rows than
\code{ann_index_flow}. The flow data in \code{mon_flow} should also contain values
for the same years as \code{ann_index_flow}, though there are no checks
performed to check this, since this is expected to be a dimensionless
matrix.}

\item{nsim}{Number of times to repeat the space/time disaggregation.}

\item{sf_sites}{The site numbers (column indeces), that will scale the
index year's volume based on the annual flow being disaggregated. The
remaining sites will select the index year directly. See \strong{Details}.}

\item{ofolder}{Optional. If specified, the disaggregated flow data and the
selected index years are saved to this folder as csv files. There will be
one csv file for each time the disaggregation is repeated (\code{nsim}). This
file will contain one column for each site (\code{nsite}), and one row for each
month of data (12 * number of years in \code{ann_flow}).}

\item{index_years}{Optional. If specified, these index years will be used
instead of selecting years based on weighted sampling via
\code{\link[=knn_get_index_year]{knn_get_index_year()}}.}

\item{k_weights}{A \code{\link[=knn_params]{knn_params()}} object. By default, it uses
\code{\link[=knn_params_default]{knn_params_default()}}.}
}
\value{
A list with two entries: \code{disagg_flow} and \code{index_years}.
\code{index_years} contains a vector of the years that were selected as index
years for the flow values from \code{ann_flow}. \code{disagg_flow} is a list, with
one entry for each simulation (\code{nsim}). Each entry is a matrix with the
same number of columns as \code{mon_flow}, and 12 * the number of rows in
\code{ann_flow}, number of rows.
}
\description{
\code{knn_space_time_disagg()} disaggregates annual flow data spatially and
temporaly (to monthly), using a spatial and monthly flow pattern selected
from an "index year". The index year is selected using a k nearest-neighbor
approach, (Knowak et al., 2010).
}
\details{
The method is described in detail in \emph{Knowak et al.} (2010). The methodology
disaggregates annual flow data (\code{ann_flow}) by selecting an index year from
\code{ann_index_flow} using \code{\link[=knn_get_index_year]{knn_get_index_year()}}. After the index year is
selected, values from \code{ann_flow} are disaggregated spatially, and temporaly
based on\code{mon_flow}. The spatial pattern is reflected by including different
sites as columns in \code{mon_flow}, and the monthly disaggregation, uses the
monthly pattern in \code{mon_flow} to disaggregate the data temporarly.
Summability is preserved using this method, if the values selected in
\code{mon_flow} are scaled and if the columns (or a subset of columns) in
\code{mon_flow} sum together to equal \code{ann_index_flow}.

In some cases, it is desirable to select monthly flow directly, instead of
scaling it. This can be performed by only scaling certain sites, using
\code{sf_sites}. \code{sf_sites} should be a boolean, or a vector of numerics. If
\code{TRUE}, then all sites are scaled. If \code{FALSE}, all sites monthly values
are selected directly. Otherwise, \code{sf_sites} should be a vector of the sites
that should be scaled, based on their column index from \code{mon_flow}. For
example, if \code{mon_flow} is a matrix with 4 columns, then setting \code{mon_flow} to
\code{c(2, 3)} will scale the values in sites 2 and 3 (columns 2 and 3), while
selecting flow values directly in sites 1 and 2.
}
\examples{

# a sample of three years of flow data
flow_mat <- cbind(c(2000, 2001, 2002), c(1400, 1567, 1325))
# made up historical data to use as index years
ind_flow <- cbind(1901:1980, rnorm(80, mean = 1500, sd = 300))
# make up monthly flow for two sites
mon_flow <- cbind(
  rnorm(80 * 12, mean = 20, sd = 5),
  rnorm(80 * 12, mean = 120, sd = 45)
)
knn_space_time_disagg(flow_mat, ind_flow, mon_flow, sf_sites = 1:2)

}
\references{
Nowak, K., Prairie, J., Rajagopalan, B., Lall, U. (2010).
A nonparametric stochastic approach for multisite disaggregation of
annual to daily streamflow. \emph{Water Resources Research.}
}
\seealso{
\code{\link[=knn_get_index_year]{knn_get_index_year()}}
}
\author{
Ken Nowak
}
