% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnstdisagg.R
\docType{package}
\name{knnstdisagg-package}
\alias{knnstdisagg}
\alias{knnstdisagg-package}
\title{knnstdisagg: Nonparametric Space-Time Streamflow Disaggregation Using KNN}
\description{
Perform space and time (annual to monthly) disaggregation of 
    streamflow using a K-nearest neighbor (knn) approach.
}
\section{Acknowledgements}{

This package implements the methods developed by Nowak et al.
}

\section{Reference}{

Nowak, K., J. Prairie, B. Rajagopalan, and U. Lall (2010), A nonparametric
stochastic approach for multisite disaggregation of annual to daily
streamflow, Water Resour. Res., 46, W08529,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2009WR008530}{doi:10.1029/2009WR008530.}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rabutler-usbr/knnstdisagg}
  \item Report bugs at \url{https://github.com/rabutler-usbr/knnstdisagg/issues}
}

}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

}
