% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnstdisagg.R
\docType{package}
\name{knnstdisagg-package}
\alias{knnstdisagg}
\alias{knnstdisagg-package}
\title{knnstdisagg: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

}
\keyword{internal}
