% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knnst.R
\name{plot.knnst}
\alias{plot.knnst}
\title{Plot resulting statistics for knn space-time disaggregation}
\usage{
\method{plot}{knnst}(
  x,
  site = "S1",
  base_units = NULL,
  which = c(13, 14, 15),
  show = FALSE,
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{knnst}}

\item{site}{The site to plot. Site name as a character.}

\item{base_units}{The (input) units of the flow that was disaggregated. These
units will be shown as the y-axis label of the plot.}

\item{which}{The subset of plots to create; specify a subset of the number
\code{1:15}. See 'Details' for the different plots.}

\item{show}{Boolean that determines if diagnostic plots are shown or only
returned. If \code{TRUE} and in interactive mode, will be able to enter through
each of the 1-15 plots.}

\item{...}{Arguments to be passed to subsequent methods.}
}
\description{
\code{plot.knnst()} implements the \code{\link[=plot]{plot()}} method for \code{knnst} objects, and relies
on ggplot2. It plots the key statistics the knn space-time disaggregation
method should be preserving.
}
\details{
\code{which} controls the plots that are created. There are both monthly and
annual plots and two main plot types: a cdf of the flows accross all of the
different disaggregations, and a panel showing the mean, max, min, variance,
lag-1 correlation, and skew accross all of the different disaggregations. The
numbers correspond to:
\itemize{
\item \code{1:12}: the different monthly cdfs.
\item \code{13}: annual cdf
\item \code{14}: monthly statistics panel
\item \code{15}: annual statistics panel
}

\code{...} is passed to \code{\link[=geom_point]{geom_point()}} and \code{\link[=labs]{labs()}}.
}
